Bug Fix
If the player has 1 health point and takes 2 damage point, the player's health is equal to 255

+++ Assets/Scripts/Enemy.cs
@@ -10,7 +10,7 @@ public class Enemy : MonoBehaviour
     private float speed;
     private Player playerScript;
 
-    public byte damage;
+    public int damage;
     // Start is called before the first frame update
     void Start()
     {
